app/requirements.txt
flask
prometheus-flask-exporter
gunicorn
kafka-python
prometheus-client
python-dotenv